feat: comment enum strings only when no enum
When the --enum is used, there will be enum types declared. The
return will be of the type of the declared enum. Then there
should be no comment specifying the possible values as it if
was strings.
However, then enums are not used, the @enum [possible_values..]
will be on the comment, specifying which string will come when
that attribute is queried. Note that when one does not specifies
the --enum, the "enum" attributes becomes strings.

Change-Id: I65453af113f9d6e1ab7e94d6afdc87b401f7ec24